refactor: Preview를 위한 DEBUG 정리
- //MARK: - Preview 구분 주석 추가
- ContentView 에서 DEBUG 용도 코드 작성
