+++ CasualConversation/Targets/Presentation/Sources/NoteSetView/NoteDetailView/NoteDetailView.swift
@@ -142,6 +142,45 @@ extension NoteDetailView {
 	
 }
 
+// MARK: - Preview
+
+#if DEBUG
+
+extension NoteDetailViewModel {
+	
+	static var previewViewModels: [NoteDetailViewModel] {
+		[
+			.init(dependency: .init(
+				useCase: DebugNoteUseCase(),
+				item: .init(
+						id: .init(),
+						original: "Way out",
+						translation: "나가는 길",
+						category: .vocabulary,
+						references: [],
+						createdDate: Date()
+					)
+				)
+			),
+			.init(dependency: .init(
+				useCase: DebugNoteUseCase(),
+				item: .init(
+						id: .init(),
+						original: "Hi, I'm Marco.\nI'm glad meet you.\nI'd like to talk to you.",
+						translation: "안녕하세요, 저는 마르코입니다.\n만나서 반갑습니다.\n이야기하기를 바랬어요.",
+						category: .sentece,
+						references: [],
+						createdDate: Date()
+					)
+				)
+			)
+		]
+	}
+	
+}
+
+#endif
+
 struct NoteDetail_Previews: PreviewProvider {
 	
 	static let viewModels = NoteDetailViewModel.previewViewModels
