+++ CasualConversation/Targets/Presentation/Sources/NoteSetView/NoteDetailView/NoteDetailViewModel.swift
@@ -46,60 +46,3 @@ extension NoteDetailViewModel {
 	}
 	
 }
-
-#if DEBUG
-
-struct DebugNoteUseCase: NoteManagable {
-	
-	func list() -> [Note] {
-		[]
-	}
-	
-	func add(item: Note, completion: (CCError?) -> Void) {
-		
-	}
-	
-	func edit(newItem: Note, completion: (CCError?) -> Void) {
-		
-	}
-	
-	func delete(item: Note, completion: (CCError?) -> Void) {
-		
-	}
-
-}
-
-extension NoteDetailViewModel {
-	
-	static var previewViewModels: [NoteDetailViewModel] {
-		[
-			.init(dependency: .init(
-				useCase: DebugNoteUseCase(),
-				item: .init(
-						id: .init(),
-						original: "Way out",
-						translation: "나가는 길",
-						category: .vocabulary,
-						references: [],
-						createdDate: Date()
-					)
-				)
-			),
-			.init(dependency: .init(
-				useCase: DebugNoteUseCase(),
-				item: .init(
-						id: .init(),
-						original: "Hi, I'm Marco.\nI'm glad meet you.\nI'd like to talk to you.",
-						translation: "안녕하세요, 저는 마르코입니다.\n만나서 반갑습니다.\n이야기하기를 바랬어요.",
-						category: .sentece,
-						references: [],
-						createdDate: Date()
-					)
-				)
-			)
-		]
-	}
-	
-}
-
-#endif
